+++ src/main/java/bleuauction/bleuauction_be/server/orderMenu/controller/OrderMenuController.java
@@ -6,7 +6,8 @@ import bleuauction.bleuauction_be.server.order.entity.Order;
 import bleuauction.bleuauction_be.server.orderMenu.dto.OrderMenuDTO;
 import bleuauction.bleuauction_be.server.orderMenu.entity.OrderMenu;
 import bleuauction.bleuauction_be.server.orderMenu.repository.OrderMenuRepository;
-import bleuauction.bleuauction_be.server.orderMenu.service.OrderMenuService;
+import bleuauction.bleuauction_be.server.orderMenu.service.OrderMenuComponentService;
+import bleuauction.bleuauction_be.server.orderMenu.service.OrderMenuModuleService;
 import bleuauction.bleuauction_be.server.common.utils.JwtUtils;
 import bleuauction.bleuauction_be.server.common.jwt.TokenMember;
 import jakarta.servlet.http.HttpSession;
@@ -24,7 +25,9 @@ import java.util.Optional;
 @RequestMapping("/api/ordermenu")
 public class OrderMenuController {
 
-  private final OrderMenuService orderMenuService;
+  private final OrderMenuComponentService orderMenuComponentService;
+  private final OrderMenuModuleService orderMenuModuleService;
+
   private final OrderMenuRepository orderMenuRepository;
   private final JwtUtils jwtUtils;
   private final MemberComponentService memberComponentService;
@@ -45,26 +48,26 @@ public class OrderMenuController {
 
     Order order = (Order) session.getAttribute("order");
 
-      return orderMenuService.addOrderMenuDTO(loginUser.get(), order, orderMenuDTO);
+      return orderMenuComponentService.addOrderMenuDTO(loginUser.get(), order, orderMenuDTO);
   }
 
 
   //주문 번호별 주문메뉴 조회, 둘 중에 하나 삭제
   @GetMapping("/{orderNo}")
   public List<OrderMenu> findOrderMenus(@PathVariable("orderNo") Long orderNo) {
-      return orderMenuService.findOrderMenusByOrderNoAndStatusY(orderNo);
+      return orderMenuComponentService.findOrderMenusByOrderNoAndStatusY(orderNo);
   }
 
   //주문별 주문 메뉴 조회
   @GetMapping("/order/{orderNo}")
   public List<OrderMenu> findOrderMenuDTOs(@PathVariable("orderNo") Long orderNo) {
-      return orderMenuService.findOrderMenuDTOsByOrderNo(orderNo);
+      return orderMenuComponentService.findOrderMenuDTOsByOrderNo(orderNo);
   }
 
   // 삭제
   @DeleteMapping("/{orderMenuNo}")
   public ResponseEntity<String> deleteOrderMenu(@PathVariable("orderMenuNo") Long orderMenuNo) {
-    orderMenuService.deleteOrderMenu(orderMenuNo);
+    orderMenuModuleService.deleteOrderMenu(orderMenuNo);
     return ResponseEntity.ok("OrderMenu deleted successfully");
   }
 
@@ -78,7 +81,7 @@ public class OrderMenuController {
 
   @PutMapping("/update/{orderMenuNo}")
   public ResponseEntity<String> updateOM (OrderMenu orderMenu, @PathVariable("orderMenuNo") Long orderMenuNo) {
-    orderMenuService.update(orderMenu);
+    orderMenuComponentService.update(orderMenu);
     log.info("ordermenu/update");
     return ResponseEntity.ok("OrderMenu updated successfully");
   }
